MCR contract changes - pack storage variables - rename `getAllSumAssurance` to `getTotalActiveCoverAmount` - sort imports

+++ contracts/interfaces/IMCR.sol
@@ -8,16 +8,16 @@ interface IMCR {
 
   function getMCR() external view returns (uint);
 
-  function mcr() external view returns (uint112);
+  function mcr() external view returns (uint80);
 
-  function desiredMCR() external view returns (uint112);
+  function desiredMCR() external view returns (uint80);
 
   function lastUpdateTime() external view returns (uint32);
 
-  function maxMCRIncrement() external view returns (uint24);
+  function maxMCRIncrement() external view returns (uint16);
 
   function gearingFactor() external view returns (uint24);
 
-  function minUpdateTime() external view returns (uint24);
+  function minUpdateTime() external view returns (uint16);
 
 }
